requirements.txt
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2024.2.0
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
monai==1.3.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.2.0
typing_extensions==4.9.0
tzdata==2023.3
zipp==3.17.0
